revolut/overlay/app/requirements.txt
python-dotenv==1.0.0
PyJWT==2.8.0
fastapi==0.104.1
mysql-connector-python==8.2.0
gunicorn==21.2.0
requests==2.31.0
uvicorn==0.24.0
cryptography==41.0.5
